Fix Picker not providing the correct sticker after searching
Picking a sticker after using the search bar returns the index of that
sticker inside the filtered list and not the original list which results
in unexpected behaviors.

Instead of filtering the sticker list according to the search prompt,
hide the unwanted stickers by not rendering them. By doing so, all stickers
will retain their original index from the master sticker list.